+++ scripts/mod_01_x=t_v=E_Ns.py
@@ -9,20 +9,20 @@ from qom.ui.plotters import MPLPlotter
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'entan_oem_mod')))
 # import system
-from systems.Mod00 import Mod00
+from systems.Mod01 import Mod01
 
 # all parameters
 params = {
     'solver': {
         'show_progress': True,
-        'cache': False,
-        'cache_dir': 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_00/0.0_1000.0_10001',
+        'cache': True,
+        'cache_dir': 'H:/Workspace/VSCode/Python/entan_oem_mod/data/mod_01/0.0_1000.0_10001',
         'method': 'ode',
         'measure_type': 'qcm',
         'qcm_type': 'entan',
         'idx_mode_i': 1,
         'idx_mode_j': 2,
-        'range_min': 0,
+        'range_min': 9371,
         'range_max': 10001,
         't_min': 0,
         't_max': 1000,
@@ -30,7 +30,7 @@ params = {
     },
     'system': {
         'A_ls': [50, 0.0],
-        'A_vs': [1e2, 0.0], 
+        'A_vs': [1e2, 10.0], 
         'Delta_0': 1.0,
         'gammas': [5e-3, 5e-2],
         'gs': [5e-3, 5e-4],
@@ -38,17 +38,20 @@ params = {
         'n_ths': [0, 0],
         'Omegas': [2.0, 2.0],
         'omegas': [1.0, 1.0],
+        'thetas': [0.1, 2.0],
         't_mod': 'cos',
         't_pos': 'bottom'
     },
     'plotter': {
         'type': 'lines',
         'show_legend': True,
-        'title': '$\\Omega_{l} = 2.0, \\Omega_{v} = 2.0$',
+        'title': '$\\theta_{1} = 2.0$',
         'x_label': '$\\omega_{0} t$',
-        'y_legend': ['$A_{l}^{(1)} = 0.0, A_{v}^{(1)} = 0.0$', '$A_{l}^{(1)} = 3.0, A_{v}^{(1)} = 1.5$'],
+        'x_ticks': [950, 960, 970, 980, 990, 1000],
+        'y_legend': ['$\\theta_{0} = 0.0$', '$\\theta_{0} = 0.1$'],
         'y_colors': ['b', 'r'],
-        'v_label': '$E_{N}$'
+        'v_label': '$E_{N}$',
+        'v_ticks': [0.175, 0.180, 0.185, 0.190]
     }
 }
 
@@ -56,16 +59,16 @@ params = {
 init_log()
 
 # initialize system
-system = Mod00(params['system'])
+system = Mod01(params['system'])
 
 # get entanglement without modulation
-system.params['A_ls'][1] = 0.0
 system.params['A_vs'][1] = 0.0
+system.params['thetas'][0] = 0.0
 M_0, T = system.get_measure_dynamics(params['solver'], system.ode_func, system.get_ivc)
 
 # get entanglement with modulation
-system.params['A_ls'][1] = 5.0
 system.params['A_vs'][1] = 10.0
+system.params['thetas'][0] = 0.1
 M_1, T = system.get_measure_dynamics(params['solver'], system.ode_func, system.get_ivc)
 
 # plotter